Refactor Policies.jsx and fix AI Search API endpoint
